+++ src/test/java/org/opensearch/knn/index/codec/jvector/KNNJVectorTests.java
@@ -1781,8 +1781,7 @@ public class KNNJVectorTests extends LuceneTestCase {
             filterQuery,
             overQueryFactor,
             KNNConstants.DEFAULT_QUERY_SIMILARITY_THRESHOLD.floatValue(),
-            KNNConstants.DEFAULT_QUERY_RERANK_FLOOR.floatValue(),
-            KNNConstants.DEFAULT_QUERY_USE_PRUNING
+            KNNConstants.DEFAULT_QUERY_RERANK_FLOOR.floatValue()
         );
     }